Updated scripts to dconf and palettes to 18 color scheme.
The new release of guake uses GTK3, for which dconf should be used if
we're not using the standard libraries.

In addition, the foreground and background colors that used to be
assigned to fonts in the appropriate section are now part of the color
scheme as the 17th (FG) and 18th (BG) colors.